+++ lib/constants.dart
@@ -1,2 +1,5 @@
+import 'package:anagrammatic/anagram.dart';
+
+// TODO: Split these into own file
 final int minimumAnagramLength = 1;
 final int maximumAnagramLength = 45;